Broke the "Libraries to Add to build path" group into a separate class
