+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/src/com/google/cloud/tools/eclipse/appengine/newproject/AppEngineStandardWizardPage.java
@@ -17,40 +17,22 @@
 package com.google.cloud.tools.eclipse.appengine.newproject;
 
 import java.io.File;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.LinkedList;
 import java.util.List;
 
-import org.eclipse.core.databinding.DataBindingContext;
-import org.eclipse.core.databinding.UpdateValueStrategy;
-import org.eclipse.core.databinding.beans.PojoProperties;
-import org.eclipse.core.databinding.observable.Realm;
-import org.eclipse.core.databinding.observable.value.IObservableValue;
-import org.eclipse.core.databinding.observable.value.WritableValue;
 import org.eclipse.core.runtime.IStatus;
-import org.eclipse.jface.databinding.swt.DisplayRealm;
-import org.eclipse.jface.databinding.swt.ISWTObservableValue;
-import org.eclipse.jface.databinding.swt.WidgetProperties;
 import org.eclipse.jface.dialogs.Dialog;
-import org.eclipse.jface.layout.GridDataFactory;
-import org.eclipse.jface.layout.GridLayoutFactory;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.ModifyEvent;
 import org.eclipse.swt.events.ModifyListener;
 import org.eclipse.swt.layout.GridData;
-import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.Group;
 import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.Text;
 import org.eclipse.ui.dialogs.WizardNewProjectCreationPage;
 import com.google.cloud.tools.eclipse.appengine.libraries.model.Library;
 import com.google.cloud.tools.eclipse.appengine.ui.AppEngineImages;
-import com.google.cloud.tools.eclipse.ui.util.databinding.BooleanConverter;
 import com.google.cloud.tools.eclipse.usagetracker.AnalyticsEvents;
 import com.google.cloud.tools.eclipse.usagetracker.AnalyticsPingManager;
-import com.google.common.base.Strings;
 
 /**
  * UI to collect all information necessary to create a new App Engine Standard Java Eclipse project.
@@ -58,9 +40,7 @@ import com.google.common.base.Strings;
 public class AppEngineStandardWizardPage extends WizardNewProjectCreationPage {
 
   private Text javaPackageField;
-  private Group apiGroup;
-  private List<Button> libraryButtons = new LinkedList<>();
-  private DataBindingContext bindingContext;
+  private AddManageLibrariesGroup addManageLibrariesGroup;
 
   public AppEngineStandardWizardPage() {
     super("basicNewProjectPage"); //$NON-NLS-1$
@@ -95,91 +75,9 @@ public class AppEngineStandardWizardPage extends WizardNewProjectCreationPage {
     javaPackageField.addModifyListener(pageValidator);
 
     // Manage APIs
-    addManageLibrariesWidgets(container);
+    addManageLibrariesGroup = new AddManageLibrariesGroup(this, container, SWT.NONE);
 
     Dialog.applyDialogFont(container);
-  }
-
-  private void addManageLibrariesWidgets(Composite container) {
-    apiGroup = new Group(container, SWT.NONE);
-    apiGroup.setText(Messages.AppEngineStandardWizardPage_librariesGroupLabel);
-    GridDataFactory.fillDefaults().span(2, 1).applyTo(apiGroup);
-
-    List<Library> libraries = getLibraries();
-    for (Library library : libraries) {
-      Button libraryButton = new Button(apiGroup, SWT.CHECK);
-      libraryButton.setText(getLibraryName(library));
-      libraryButton.setData(library);
-      libraryButtons.add(libraryButton);
-    }
-
-    addDatabindingForDependencies();
-
-    GridLayoutFactory.fillDefaults().applyTo(apiGroup);
-  }
-
-  private static String getLibraryName(Library library) {
-    if (!Strings.isNullOrEmpty(library.getName())) {
-      return library.getName();
-    } else {
-      return library.getId();
-    }
-  }
-
-  private void addDatabindingForDependencies() {
-    bindingContext = new DataBindingContext();
-    for (Button libraryButton : libraryButtons) {
-      Library library = (Library) libraryButton.getData();
-      if (!library.getLibraryDependencies().isEmpty()) {
-        addDatabindingForDependencies(libraryButton);
-      }
-    }
-  }
-
-  private void addDatabindingForDependencies(Button libraryButton) {
-    Library library = (Library) libraryButton.getData();
-    for (String libraryId : library.getLibraryDependencies()) {
-      Button dependencyButton = getButtonForLibraryId(libraryId);
-      if (dependencyButton != null) {
-        ISWTObservableValue libraryButtonSelection = WidgetProperties.selection().observe(libraryButton);
-        IObservableValue dependencyButtonSelection =
-            PojoProperties.value(Button.class, "selection").observe(getDisplayRealm(), dependencyButton);
-        IObservableValue dependencyButtonEnablement =
-            PojoProperties.value(Button.class, "enabled").observe(getDisplayRealm(), dependencyButton);
-
-        WritableValue intermediate = new WritableValue(false, Boolean.class);
-        bindingContext.bindValue(libraryButtonSelection, intermediate);
-        bindingContext.bindValue(dependencyButtonSelection, intermediate);
-        bindingContext.bindValue(dependencyButtonEnablement, intermediate,
-                                 new UpdateValueStrategy().setConverter(BooleanConverter.negate()),
-                                 new UpdateValueStrategy().setConverter(BooleanConverter.negate()));
-      }
-    }
-  }
-
-  private Realm getDisplayRealm() {
-    return DisplayRealm.getRealm(getControl().getDisplay());
-  }
-
-  private Button getButtonForLibraryId(String libraryId) {
-    for (Button button : libraryButtons) {
-      Library library = (Library) button.getData();
-      if (library.getId().equals(libraryId)) {
-        return button;
-      }
-    }
-    return null;
-  }
-
-  // TODO obtain libraries from extension registry
-  // https://github.com/GoogleCloudPlatform/google-cloud-eclipse/issues/819
-  private List<Library> getLibraries() {
-    Library appEngine = new Library("appengine-api");
-    appEngine.setName("App Engine API");
-    Library endpoints = new Library("appengine-endpoints");
-    endpoints.setName("App Engine Endpoints");
-    endpoints.setLibraryDependencies(Collections.singletonList("appengine-api"));
-    return Arrays.asList(appEngine, endpoints);
   }
 
   @Override
@@ -221,20 +119,12 @@ public class AppEngineStandardWizardPage extends WizardNewProjectCreationPage {
   }
 
   public List<Library> getSelectedLibraries() {
-    List<Library> selected = new LinkedList<>();
-    for (Button button : libraryButtons) {
-      if (button.getSelection()) {
-        selected.add((Library) button.getData());
-      }
-    }
-    return selected;
+    return addManageLibrariesGroup.getSelectedLibraries();
   }
 
   @Override
   public void dispose() {
-    if (bindingContext != null) {
-      bindingContext.dispose();
-    }
+    addManageLibrariesGroup.dispose();
     super.dispose();
   }
 }
